% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp.R
\name{create_rtemp}
\alias{create_rtemp}
\title{Create a new directory with the specified R template}
\usage{
create_rtemp(dirname = "new-dir", template = "united_html")
}
\arguments{
\item{dirname}{name of the directory to create. If not specified, a directory
\emph{new-dir} will be created.}

\item{template}{which template to use?}
}
\description{
This function creates a new directory under the current directory, which will
contain the given files of the specified template, ready to be rendered.
}
\details{
The available templates are: \emph{united_html} (default).
The name of the main \code{.Rmd} file will be \emph{test.Rmd}.
}
\examples{
\dontrun{
create_rtemp(dirname = "tmp_dir", format = "united_html")
}

}
